+++ ricaun.RevitTest.Console/Revit/Utils/VisualStudioDebugUtils.cs
@@ -1,5 +1,5 @@
 ﻿using EnvDTE;
-using ricaun.RevitTest.Console.Utils;
+using ricaun.RevitTest.Command.Utils;
 using System.Diagnostics;
 using System.Linq;