+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.8] / 2023-06-13
+### TestAdapter
+- Update to show OutputConsole in the Log.
+- Update `TestModels`, `TestAssemblyModel` and `TestCaseModel`
+- Update `TestAssemblyModel` to record TestCase with the method `RecordResultTestModel`
+- Fix Download Application File
+
 ## [1.0.7] / 2023-06-12
 ### Console
 - Decouple using interface `IRunTestService`
@@ -137,6 +144,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [x] TestsFail
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.0.8]: ../../compare/1.0.7...1.0.8
 [1.0.7]: ../../compare/1.0.6...1.0.7
 [1.0.6]: ../../compare/1.0.5...1.0.6
 [1.0.5]: ../../compare/1.0.4...1.0.5